Refactor: created moved heatmap logic to corresponding header file
